+++ src/routes/mod.ts
@@ -11,9 +11,12 @@ const router = express.Router();
 router.delete('/clip/:id', isAuth, isMod, async (req, res) => {
   const { id } = req.params;
   try {
+    if (isNaN(id as any)) throw Error("id must be a number");
     await Clip.query().deleteById(id);
+    // TODO: Delete tags
     return res.status(200).send();
   } catch (error) {
+    console.error(error.message);
     return res.status(400).send();
   }
 });
